Clean old classes
