+++ Project.swift
@@ -185,7 +185,6 @@ let project = Project(name: "kDrive",
                 "kDrive/UI/Controller/Files/Rights and Share/**",
                 "kDrive/UI/Controller/Files/Save File/**",
                 "kDrive/UI/Controller/Files/Search/**",
-                "kDrive/UI/Controller/Files/FileListCollectionViewController.swift",
                 "kDrive/UI/Controller/Files/FileListViewController.swift",
                 "kDrive/UI/Controller/Files/FloatingPanelSortOptionTableViewController.swift",
                 "kDrive/UI/Controller/Floating Panel Information/**",
@@ -279,7 +278,6 @@ let project = Project(name: "kDrive",
                 "kDrive/UI/Controller/Files/Rights and Share/**",
                 "kDrive/UI/Controller/Files/Save File/**",
                 "kDrive/UI/Controller/Files/Search/**",
-                "kDrive/UI/Controller/Files/FileListCollectionViewController.swift",
                 "kDrive/UI/Controller/Files/FileListViewController.swift",
                 "kDrive/UI/Controller/Files/FloatingPanelSortOptionTableViewController.swift",
                 "kDrive/UI/Controller/Floating Panel Information/**",
