+++ kDrive/UI/Controller/Files/FileListViewController.swift
@@ -823,8 +823,8 @@ extension FileListViewController: FilesHeaderViewDelegate {
         }
 
         func moveButtonPressed() {
-            let selectFolderNavigationController = SelectFolderViewController2.instantiateInNavigationController(driveFileManager: driveFileManager)
-            let selectFolderViewController = selectFolderNavigationController.topViewController as? SelectFolderViewController2
+            let selectFolderNavigationController = SelectFolderViewController.instantiateInNavigationController(driveFileManager: driveFileManager)
+            let selectFolderViewController = selectFolderNavigationController.topViewController as? SelectFolderViewController
             selectFolderViewController?.disabledDirectoriesSelection = [selectedFiles.first?.parent ?? driveFileManager.getRootFile()]
             selectFolderViewController?.selectHandler = { selectedFolder in
                 let group = DispatchGroup()
