+++ kDrive/UI/Controller/Files/Save File/SelectFolderViewController.swift
@@ -24,18 +24,13 @@ protocol SelectFolderDelegate: AnyObject {
     func didSelectFolder(_ folder: File)
 }
 
-class SelectFolderViewController: FileListCollectionViewController {
+class SelectFolderViewController: FileListViewController {
+
+    override class var storyboard: UIStoryboard { Storyboard.saveFile }
+    override class var storyboardIdentifier: String { "SelectFolderViewController" }
 
     @IBOutlet weak var selectFolderButton: UIButton!
     @IBOutlet weak var addFolderButton: UIBarButtonItem!
-
-    override var isMultipleSelectionEnabled: Bool {
-        return false
-    }
-
-    override var showUploadingFiles: Bool {
-        return false
-    }
 
     var disabledDirectoriesSelection = [File]()
     var fileToMove: Int? = nil
@@ -51,6 +46,9 @@ class SelectFolderViewController: FileListCollectionViewController {
     }
 
     override func viewDidLoad() {
+        // Set configuration
+        configuration = Configuration(showUploadingFiles: false, isMultipleSelectionEnabled: false, rootTitle: KDriveStrings.Localizable.selectFolderTitle, emptyViewType: .emptyFolder)
+
         super.viewDidLoad()
 
         collectionView.contentInset = UIEdgeInsets(top: 0, left: 0, bottom: UIConstants.listFloatingButtonPaddingBottom, right: 0)
@@ -62,51 +60,36 @@ class SelectFolderViewController: FileListCollectionViewController {
         addFolderButton.accessibilityLabel = KDriveStrings.Localizable.createFolderTitle
         selectFolderButton.isEnabled = !(disabledDirectoriesSelection.map(\.id).contains(currentDirectory.id)) && (currentDirectory.rights?.moveInto.value ?? false || currentDirectory.rights?.createNewFile.value ?? false)
         if currentDirectory.id == DriveFileManager.constants.rootID {
-            // Root directory
-            navigationItem.title = KDriveStrings.Localizable.selectFolderTitle
-            // Set back button if the view controller is presented modally
-            let nbViewControllers = navigationController?.viewControllers.count ?? 0
-            if isModal && nbViewControllers < 2 {
-                navigationItem.leftBarButtonItem = UIBarButtonItem(barButtonSystemItem: .stop, target: self, action: #selector(close))
+            // Root directory: set back button if the view controller is presented modally
+            let viewControllersCount = navigationController?.viewControllers.count ?? 0
+            if isModal && viewControllersCount < 2 {
+                navigationItem.leftBarButtonItem = UIBarButtonItem(barButtonSystemItem: .stop, target: self, action: #selector(closeButtonPressed))
                 navigationItem.leftBarButtonItem?.accessibilityLabel = KDriveStrings.Localizable.buttonClose
             }
         }
     }
 
-    override func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let file = sortedChildren[indexPath.row]
-        let cell = super.collectionView(collectionView, cellForItemAt: indexPath) as! FileCollectionViewCell
-        cell.setEnabled(file.isDirectory && file.id != fileToMove)
-        cell.moreButton.isHidden = true
-        return cell
+    static func instantiateInNavigationController(driveFileManager: DriveFileManager) -> TitleSizeAdjustingNavigationController {
+        let selectFolderViewController = instantiate(driveFileManager: driveFileManager)
+        let navigationController = TitleSizeAdjustingNavigationController(rootViewController: selectFolderViewController)
+        navigationController.navigationBar.prefersLargeTitles = true
+        return navigationController
     }
 
-    override func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        let selectedFile = sortedChildren[indexPath.row]
-        if selectedFile.isDirectory {
-            let nextVC = SelectFolderViewController.instantiate(driveFileManager: driveFileManager)
-            nextVC.disabledDirectoriesSelection = disabledDirectoriesSelection
-            nextVC.fileToMove = fileToMove
-            nextVC.currentDirectory = selectedFile
-            nextVC.delegate = delegate
-            nextVC.selectHandler = selectHandler
-            nextVC.sortType = sortType
-            navigationController?.pushViewController(nextVC, animated: true)
-        }
-    }
+    // MARK: - Actions
 
-    @objc func close() {
+    @objc func closeButtonPressed() {
         dismiss(animated: true)
     }
 
     @IBAction func selectButtonPressed(_ sender: UIButton) {
         delegate?.didSelectFolder(currentDirectory)
         selectHandler?(currentDirectory)
-        //We are only selecting files we can dismiss
+        // We are only selecting files we can dismiss
         if navigationController?.viewControllers.first is SelectFolderViewController {
             navigationController?.dismiss(animated: true)
         } else {
-            //We are creating file, go back to file name
+            // We are creating file, go back to file name
             navigationController?.popToRootViewController(animated: true)
         }
     }
@@ -116,17 +99,29 @@ class SelectFolderViewController: FileListCollectionViewController {
         navigationController?.present(newFolderViewController, animated: true)
     }
 
-    class func instantiateInNavigationController(driveFileManager: DriveFileManager) -> TitleSizeAdjustingNavigationController {
-        let selectFolderViewController = instantiate(driveFileManager: driveFileManager)
-        let navigationController = TitleSizeAdjustingNavigationController(rootViewController: selectFolderViewController)
-        navigationController.navigationBar.prefersLargeTitles = true
-        return navigationController
+    // MARK: - Collection view data source
+
+    override func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        let file = sortedFiles[indexPath.row]
+        let cell = super.collectionView(collectionView, cellForItemAt: indexPath) as! FileCollectionViewCell
+        cell.setEnabled(file.isDirectory && file.id != fileToMove)
+        cell.moreButton.isHidden = true
+        return cell
     }
 
-    override class func instantiate(driveFileManager: DriveFileManager) -> SelectFolderViewController {
-        let viewController = UIStoryboard(name: "SaveFile", bundle: nil).instantiateViewController(withIdentifier: "SelectFolderViewController") as! SelectFolderViewController
-        viewController.driveFileManager = driveFileManager
-        return viewController
+    // MARK: - Collection view delegate
+
+    override func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+        let selectedFile = sortedFiles[indexPath.row]
+        if selectedFile.isDirectory {
+            let nextVC = SelectFolderViewController.instantiate(driveFileManager: driveFileManager)
+            nextVC.disabledDirectoriesSelection = disabledDirectoriesSelection
+            nextVC.fileToMove = fileToMove
+            nextVC.currentDirectory = selectedFile
+            nextVC.delegate = delegate
+            nextVC.selectHandler = selectHandler
+            navigationController?.pushViewController(nextVC, animated: true)
+        }
     }
 
     // MARK: - State restoration
@@ -147,4 +142,5 @@ class SelectFolderViewController: FileListCollectionViewController {
         }
         setUpDirectory()
     }
+
 }
