+++ kDrive/UI/Controller/Files/Search/SearchViewController.swift
@@ -19,6 +19,42 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 import UIKit
 import kDriveCore
 
+public struct FileTypeRow: RawRepresentable {
+
+    public typealias RawValue = String
+
+    public var rawValue: String
+    let name: String
+    let icon: UIImage
+    let type: String
+
+    public init?(rawValue: String) {
+        if let type = FileTypeRow.allValues.first(where: { $0.rawValue == rawValue }) {
+            self = type
+        } else {
+            return nil
+        }
+    }
+
+    init(rawValue: String, name: String, icon: UIImage, type: String) {
+        self.rawValue = rawValue
+        self.name = name
+        self.icon = icon
+        self.type = type
+    }
+
+    static let imagesRow = FileTypeRow(rawValue: "images", name: KDriveStrings.Localizable.allPictures, icon: KDriveAsset.fileImage.image, type: ConvertedType.image.rawValue)
+    static let videosRow = FileTypeRow(rawValue: "videos", name: KDriveStrings.Localizable.allVideo, icon: KDriveAsset.fileVideo.image, type: ConvertedType.video.rawValue)
+    static let audioRow = FileTypeRow(rawValue: "audio", name: KDriveStrings.Localizable.allAudio, icon: KDriveAsset.fileAudio.image, type: ConvertedType.audio.rawValue)
+    static let pdfRow = FileTypeRow(rawValue: "pdf", name: KDriveStrings.Localizable.allPdf, icon: KDriveAsset.filePdf.image, type: ConvertedType.pdf.rawValue)
+    static let docsRow = FileTypeRow(rawValue: "docs", name: KDriveStrings.Localizable.allOfficeDocs, icon: KDriveAsset.fileText.image, type: ConvertedType.text.rawValue)
+    static let pointsRow = FileTypeRow(rawValue: "points", name: KDriveStrings.Localizable.allOfficePoints, icon: KDriveAsset.filePresentation.image, type: ConvertedType.presentation.rawValue)
+    static let gridsRow = FileTypeRow(rawValue: "grid", name: KDriveStrings.Localizable.allOfficeGrids, icon: KDriveAsset.fileSheets.image, type: ConvertedType.spreadsheet.rawValue)
+    static let folderRow = FileTypeRow(rawValue: "folder", name: KDriveStrings.Localizable.allFolder, icon: KDriveAsset.folderFilled.image, type: ConvertedType.folder.rawValue)
+    static let dropboxRow = FileTypeRow(rawValue: "dropbox", name: KDriveStrings.Localizable.dropBoxTitle, icon: KDriveAsset.folderDropBox.image, type: "")
+    static let allValues = [imagesRow, videosRow, audioRow, pdfRow, docsRow, pointsRow, gridsRow, folderRow, dropboxRow]
+}
+
 class SearchViewController: FileListViewController {
 
     override class var storyboard: UIStoryboard { Storyboard.search }
